+++ src/api/calendarApi.js
@@ -7,7 +7,7 @@ export const calendarApi = {
       const data = await request.get('/cheongsan/calendar/transactions', {
         params: { year, month },
       });
-      console.log('월별 거래내역 조회:', data);
+      // console.log('월별 거래내역 조회:', data);
       return data;
     } catch (error) {
       console.error('월별 거래내역 조회 실패:', error);
@@ -21,7 +21,7 @@ export const calendarApi = {
       const data = await request.get('/cheongsan/calendar/repayments', {
         params: { year, month },
       });
-      console.log('월별 대출 상환일자 조회:', data);
+      // console.log('월별 대출 상환일자 조회:', data);
       return data;
     } catch (error) {
       console.error('월별 대출 상환일자 조회 실패:', error);
@@ -33,7 +33,7 @@ export const calendarApi = {
   getDailyRepayments: async (date) => {
     try {
       const data = await request.get(`/cheongsan/calendar/repayments/${date}`);
-      console.log('일별 대출 상환일자 조회:', data);
+      // console.log('일별 대출 상환일자 조회:', data);
       return data;
     } catch (error) {
       console.error('일별 대출 상환일자 조회 실패:', error);
@@ -47,7 +47,7 @@ export const calendarApi = {
       const data = await request.get(
         `/cheongsan/calendar/transactions/${date}`
       );
-      console.log('일별 거래 내역 조회:', data);
+      // console.log('일별 거래 내역 조회:', data);
       return data;
     } catch (error) {
       console.error('일별 거래 내역 조회 실패:', error);
