+++ src/assets/styles/pages/Login.module.css
@@ -5,7 +5,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  background-color: var(--color-gray3);
+  background-color: white;
   padding: 20px;
 }
 
@@ -14,29 +14,27 @@
   max-width: 335px;
   display: flex;
   flex-direction: column;
+  min-height: calc(100vh - 40px); /* 패딩 20px * 2 = 40px 제외 */
 }
 
 /* 타이틀 영역 */
 .title {
   font-family: var(--font-family);
   font-weight: 600;
-  font-size: 24px;
+  font-size: 32px;
   line-height: 1.2;
   letter-spacing: -0.5px;
-  color: var(--color-main);
-  margin: 0 0 8px 0;
-  text-align: left;
+  margin-bottom: 33px;
 }
 
 .subtitle {
   font-family: var(--font-family);
   font-weight: 400;
-  font-size: 14px;
+  font-size: 16px;
   line-height: 1.4;
   letter-spacing: -0.25px;
-  color: var(--color-gray0);
-  margin: 0 0 40px 0;
   text-align: left;
+  margin-bottom: 32px;
 }
 
 /* 로그인 폼 */
@@ -44,7 +42,6 @@
   display: flex;
   flex-direction: column;
   gap: 16px;
-  margin-bottom: 32px;
 }
 
 .inputGroup {
@@ -55,13 +52,12 @@
   width: 100%;
   height: 48px;
   padding: 0 16px;
-  border: 1px solid var(--color-gray1);
-  border-radius: 8px;
+  border: 1px solid var(--color-gray0);
   background-color: white;
   font-family: var(--font-family);
   font-weight: 400;
   font-size: 16px;
-  color: var(--color-main);
+  color: var(--color-gray-unnamed);
   transition: border-color 0.2s ease;
 }
 
@@ -96,12 +92,12 @@
   width: 100%;
   height: 48px;
   background-color: var(--color-main);
-  color: white;
+  color: var(--color-gray3);
   border: none;
   border-radius: 8px;
   font-family: var(--font-family);
-  font-weight: 600;
-  font-size: 16px;
+  font-weight: 700;
+  font-size: 20px;
   letter-spacing: -0.25px;
   cursor: pointer;
   transition: all 0.2s ease;
@@ -131,74 +127,37 @@
 .divider {
   position: relative;
   text-align: center;
-  margin: 24px 0;
-}
-
-.divider::before {
-  content: '';
-  position: absolute;
-  top: 50%;
-  left: 0;
-  right: 0;
-  height: 1px;
-  background-color: var(--color-gray1);
-}
-
-.divider span {
-  background-color: var(--color-gray3);
-  padding: 0 16px;
-  font-size: 12px;
-  color: var(--color-gray0);
-  font-weight: 400;
-}
-
-/* 네이버 로그인 버튼 */
-.naverButton {
-  width: 100%;
-  height: 48px;
-  background-color: var(--color-naver);
-  color: white;
-  border: none;
-  border-radius: 8px;
-  font-family: var(--font-family);
-  font-weight: 600;
-  font-size: 16px;
-  letter-spacing: -0.25px;
-  cursor: pointer;
-  transition: all 0.2s ease;
+  margin: 32px 0 24px 0;
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 8px;
-  margin-bottom: 40px;
 }
 
-.naverButton:hover:not(:disabled) {
-  background-color: #02b351;
+.divider::before,
+.divider::after {
+  content: '';
+  flex: 1;
+  height: 1px;
+  background-color: #000;
 }
 
-.naverButton:active:not(:disabled) {
-  transform: translateY(1px);
+.divider::before {
+  margin-right: 16px;
 }
 
-.naverButton:disabled {
-  background-color: var(--color-gray1);
-  color: var(--color-gray0);
-  cursor: not-allowed;
-  transform: none;
+.divider::after {
+  margin-left: 16px;
 }
 
-.naverButton span {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 20px;
-  height: 20px;
-  background-color: white;
-  color: var(--color-naver);
-  border-radius: 2px;
-  font-weight: 700;
+.divider span {
+  flex-shrink: 0;
+  font-family: var(--font-family);
+  font-weight: 500; /* Medium */
   font-size: 14px;
+  line-height: 1.4;
+  letter-spacing: -0.25px;
+  color: #000;
+  white-space: nowrap;
 }
 
 /* 하단 링크 영역 */
@@ -207,14 +166,14 @@
   align-items: center;
   justify-content: center;
   gap: 8px;
-  margin-top: auto;
-  padding-top: 20px;
+  margin-top: auto; /* 남은 공간을 모두 차지해서 바닥으로 밀어냄 */
+  padding-top: 40px;
+  padding-bottom: 20px; /* 바닥에서 약간의 여백 */
 }
 
 .linkButton {
   background: none;
   border: none;
-  color: var(--color-gray0);
   font-size: 14px;
   font-weight: 400;
   cursor: pointer;
@@ -227,6 +186,6 @@
 }
 
 .separator {
-  color: var(--color-gray1);
-  font-size: 14px;
+  color: #000;
+  font-size: 20px;
 }
